server/requirements.txt
openenv-core
fastapi
uvicorn[standard]
pydantic>=2.0
